% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_functions.R
\name{calculate_correlation}
\alias{calculate_correlation}
\title{Calculate correlation text for annotation}
\usage{
calculate_correlation(x, y)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}
}
\value{
Formatted correlation string
}
\description{
Calculate correlation text for annotation
}
\keyword{internal}
