% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_functions.R
\name{calculateCorrelation}
\alias{calculateCorrelation}
\title{Calculate correlation text for annotation}
\usage{
calculateCorrelation(x, y)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}
}
\value{
Formatted correlation string
}
\description{
Calculate correlation text for annotation
}
\keyword{internal}
