% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{create_gwas_gene_highlight}
\alias{create_gwas_gene_highlight}
\title{Create GWAS gene highlight plot}
\usage{
create_gwas_gene_highlight(gene)
}
\arguments{
\item{gene}{Gene symbol}
}
\value{
ggplot2 object or NULL
}
\description{
Create GWAS gene highlight plot
}
\keyword{internal}
