% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{create_hla_association_plot}
\alias{create_hla_association_plot}
\title{Create HLA association plot}
\usage{
create_hla_association_plot(gene)
}
\arguments{
\item{gene}{Gene symbol}
}
\value{
ggplot2 object or NULL
}
\description{
Create HLA association plot
}
\keyword{internal}
