% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{createGwasGeneHighlight}
\alias{createGwasGeneHighlight}
\title{Create GWAS gene highlight plot}
\usage{
createGwasGeneHighlight(gene)
}
\arguments{
\item{gene}{Gene symbol}
}
\value{
ggplot2 object or NULL
}
\description{
Create GWAS gene highlight plot
}
\keyword{internal}
