% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{createHlaAssociationPlot}
\alias{createHlaAssociationPlot}
\title{Create HLA association plot}
\usage{
createHlaAssociationPlot(gene)
}
\arguments{
\item{gene}{Gene symbol}
}
\value{
ggplot2 object or NULL
}
\description{
Create HLA association plot
}
\keyword{internal}
