% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_expression}
\alias{example_expression}
\title{Example expression dataset}
\format{
A data frame with 240 rows and 4 variables:
\describe{
\item{gene}{Gene symbol}
\item{sample}{Sample identifier}
\item{condition}{Condition, either 'CeD' or 'Control'}
\item{expression}{Simulated expression value}
}
}
\usage{
example_expression
}
\description{
A simulated dataset containing gene expression data for celiac disease and control samples.
}
\examples{
data(example_expression)
head(example_expression)
}
\keyword{datasets}
