% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_expression_matrix}
\alias{example_expression_matrix}
\title{Example expression matrix}
\format{
A matrix with 12 rows (genes) and 10 columns (samples)
}
\usage{
example_expression_matrix
}
\description{
A simulated gene expression matrix for demonstration purposes.
}
\examples{
data(example_expression_matrix)
head(example_expression_matrix)
}
\keyword{datasets}
