% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_metadata}
\alias{example_metadata}
\title{Example sample metadata}
\format{
A data frame with 10 rows and 2 variables:
\describe{
\item{sample}{Sample identifier}
\item{condition}{Condition, either 'CeD' or 'Control'}
}
}
\usage{
example_metadata
}
\description{
Sample annotation data matching the example expression datasets.
}
\examples{
data(example_metadata)
head(example_metadata)
}
\keyword{datasets}
