% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{get_gene_summary}
\alias{get_gene_summary}
\title{Create summary statistics for gene}
\usage{
get_gene_summary(gene)
}
\arguments{
\item{gene}{Gene symbol}
}
\value{
Data.frame with summary stats
}
\description{
Create summary statistics for gene
}
\keyword{internal}
