% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_functions.R
\name{get_hla_allele_frequencies}
\alias{get_hla_allele_frequencies}
\title{Get example HLA allele frequency data}
\usage{
get_hla_allele_frequencies()
}
\value{
Data.frame with HLA allele frequencies
}
\description{
Get example HLA allele frequency data
}
\keyword{internal}
