% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_functions.R
\name{get_hla_expression_data}
\alias{get_hla_expression_data}
\title{Get example HLA expression data}
\usage{
get_hla_expression_data()
}
\value{
Data.frame with HLA expression data
}
\description{
Get example HLA expression data
}
\keyword{internal}
