% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{getGeneSummary}
\alias{getGeneSummary}
\title{Create summary statistics for gene}
\usage{
getGeneSummary(gene)
}
\arguments{
\item{gene}{Gene symbol}
}
\value{
Data.frame with summary stats
}
\description{
Create summary statistics for gene
}
\keyword{internal}
