% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_functions.R
\name{getHlaAlleleFrequencies}
\alias{getHlaAlleleFrequencies}
\title{Get example HLA allele frequency data}
\usage{
getHlaAlleleFrequencies()
}
\value{
Data.frame with HLA allele frequencies
}
\description{
Get example HLA allele frequency data
}
\keyword{internal}
