% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_functions.R
\name{getHlaExpressionData}
\alias{getHlaExpressionData}
\title{Get example HLA expression data}
\usage{
getHlaExpressionData()
}
\value{
Data.frame with HLA expression data
}
\description{
Get example HLA expression data
}
\keyword{internal}
