% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{getInteractingGenes}
\alias{getInteractingGenes}
\title{Get genes that interact with target gene}
\usage{
getInteractingGenes(gene)
}
\arguments{
\item{gene}{Gene symbol}
}
\value{
Vector of interacting genes
}
\description{
Get genes that interact with target gene
}
\keyword{internal}
