% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_functions.R
\name{mapToPathways}
\alias{mapToPathways}
\title{Map genes to pathways}
\usage{
mapToPathways(genes)
}
\arguments{
\item{genes}{Character vector of gene symbols}
}
\value{
Data.frame with pathway information
}
\description{
Map genes to pathways
}
