% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_functions.R
\name{plot_top_gwas_hits}
\alias{plot_top_gwas_hits}
\title{Plot top GWAS hits with gene labels}
\usage{
plot_top_gwas_hits(
  gwas_data,
  top_n = 20,
  title = "Top Celiac Disease GWAS Hits"
)
}
\arguments{
\item{gwas_data}{GWAS data from get_celiac_gwas()}

\item{top_n}{Number of top hits to display}

\item{title}{Plot title}
}
\description{
Plot top GWAS hits with gene labels
}
