% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_functions.R
\name{plotGwasSummary}
\alias{plotGwasSummary}
\title{Plot GWAS summary (volcano plot) since we lack chromosome data for Manhattan plot}
\usage{
plotGwasSummary(
  gwas_data,
  title = "Celiac Disease GWAS Summary",
  significance_level = 5e-08
)
}
\arguments{
\item{gwas_data}{GWAS data from getCeliacGwas()}

\item{title}{Plot title}

\item{significance_level}{Genome-wide significance level}
}
\description{
Plot GWAS summary (volcano plot) since we lack chromosome data for Manhattan plot
}
\examples{
\donttest{
# Get GWAS data and create summary plot
gwas_data <- getCeliacGwas()
plotGwasSummary(gwas_data)

# Custom title and significance level
plotGwasSummary(gwas_data, 
                title = "Celiac Disease Association Results",
                significance_level = 1e-6)  
}
}
\references{
Dudbridge, F., & Gusnanto, A. (2008). Estimation of significance
thresholds for genomewide association scans. \emph{Genetic Epidemiology}, \emph{32}(3),
227-234. https://doi.org/10.1002/gepi.20297
}
