% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppi_data}
\alias{ppi_data}
\title{Curated Protein-Protein Interaction Data}
\format{
A list with two components:
\describe{
\item{edges}{A data frame with interaction data:
\describe{
\item{gene1}{First interacting gene}
\item{gene2}{Second interacting gene}
\item{score}{Interaction confidence score (0-1)}
\item{evidence}{Type of evidence: complex, signaling, costimulation, predicted}
}
}
\item{nodes}{A data frame with gene information:
\describe{
\item{gene}{Gene symbol}
\item{type}{Gene type: MHC, Cytokine/Chemokine, Transcription Factor, Cell Surface, Other}
\item{importance}{Relative importance score (0-1)}
}
}
}
}
\usage{
ppi_data
}
\description{
A curated dataset containing protein-protein interactions for
celiac disease-related genes.
}
\examples{
data(ppi_data)
head(ppi_data$edges)
head(ppi_data$nodes)
}
\keyword{datasets}
