% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_metadata}
\alias{example_metadata}
\title{Example metadata}
\format{
A data frame with columns:
\describe{
\item{sample}{Sample identifier}
\item{condition}{Condition (CeD or Control)}
}
}
\source{
Simulated sample information for demonstration purposes
}
\usage{
data(example_metadata)
}
\description{
Example metadata
}
\references{
Leek, J. T., Scharpf, R. B., Bravo, H. C., Simcha, D., Langmead, B.,
Johnson, W. E., ... & Irizarry, R. A. (2010). Tackling the widespread and
critical impact of batch effects in high-throughput data. Nature Reviews
Genetics, 11(10), 733-739. https://doi.org/10.1038/nrg2825
}
\keyword{datasets}
