% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_metadata}
\alias{example_metadata}
\title{Example Sample Metadata}
\format{
A data frame with 50 rows and 3 variables:
\describe{
\item{sample_id}{Sample identifier}
\item{condition}{Disease condition}
\item{batch}{Processing batch}
}
}
\source{
Simulated sample information for demonstration purposes
}
\usage{
example_metadata
}
\description{
Simulated sample metadata matching the example expression datasets.
}
\references{
Leek, J. T., Scharpf, R. B., Bravo, H. C., Simcha, D., Langmead, B.,
Johnson, W. E., ... & Irizarry, R. A. (2010). Tackling the widespread and
critical impact of batch effects in high-throughput data. Nature Reviews
Genetics, 11(10), 733-739. https://doi.org/10.1038/nrg2825
}
\keyword{datasets}
