% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_functions.R
\name{getCeliacGwas}
\alias{getCeliacGwas}
\title{Get celiac disease GWAS data from GWAS Catalog}
\usage{
getCeliacGwas(fetch = FALSE, force_fetch = FALSE)
}
\arguments{
\item{fetch}{Logical, whether to fetch from GWAS Catalog or use bundled data}

\item{force_fetch}{Logical, whether to force fetching even if bundled data exists}
}
\value{
Data frame with GWAS data
}
\description{
Get celiac disease GWAS data from GWAS Catalog
}
