% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_functions.R
\name{map_to_pathways}
\alias{map_to_pathways}
\title{Map genes to pathways}
\usage{
map_to_pathways(genes)
}
\arguments{
\item{genes}{Character vector of gene symbols}
}
\value{
Data.frame with pathway information
}
\description{
Map genes to pathways
}
