% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_functions.R
\name{plotGwasSummary}
\alias{plotGwasSummary}
\title{Plot GWAS summary (volcano plot) since we lack chromosome data for Manhattan plot}
\usage{
plotGwasSummary(
  gwas_data,
  title = "Celiac Disease GWAS Summary",
  significance_level = 5e-08
)
}
\arguments{
\item{gwas_data}{GWAS data from getCeliacGwas()}

\item{title}{Plot title}

\item{significance_level}{Genome-wide significance level}
}
\description{
Plot GWAS summary (volcano plot) since we lack chromosome data for Manhattan plot
}
