% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_functions.R
\name{plotTopGwasHits}
\alias{plotTopGwasHits}
\title{Plot top GWAS hits with gene labels}
\usage{
plotTopGwasHits(gwas_data, top_n = 20, title = "Top Celiac Disease GWAS Hits")
}
\arguments{
\item{gwas_data}{GWAS data from getCeliacGwas()}

\item{top_n}{Number of top hits to display}

\item{title}{Plot title}
}
\description{
Plot top GWAS hits with gene labels
}
\references{
Sollid, L. M., & Jabri, B. (2013). Triggers and drivers of autoimmunity:
Lessons from coeliac disease. Nature Reviews Immunology, 13(4), 294-302.
https://doi.org/10.1038/nri3407
}
